Add interfaces.h for generating interface list (#119)
Currently, we should add a new interface into mojom based service,
we should modify many places. It's very annoying and repeated.
So, this patch introduces `interfaces.h` to export all mojoms and
utility functions, manifest and so on. After this patch, we no longer
modify many places to add a new interface.

Refs: #109