Back to using the `.patch` file for upstream changes (#218)
Introduced a new tool for upstream changes at #211, it would be nice in
specific use case. For example, it provided cleaned diff for upstream
changes.
That being said, it's very vulnerable to the addition of new change
because we have to copy an original upstream file to make a diff. It
required that developer should always be smart. It's not good. So, this
patch backs to the `.patch` file way. It still causes some
inconvenience, but we can overcome by the patch file viewer.

Refs: none