Make the initializeIfNeeded() be called on browser process only (#277)
Before this patch, whenever any process is initiated, the initialization
method is called. To avoid the duplicated calls, we had a check flag
but we are not sure if the initialization logic always run on browser.
It means that the logic might be run on renderer or other processes.

So, this patch fixes the problem. Expose the isBrowserProcess() method
in upstream side, and then use it to check whether it is actually
browser process or not.

Refs: #267, #274