Fix a bug that the blacklist is removed during password update (#287)
Currently, if users update an already stored credential on the
black-listed origin, there is a bug that the origin is removed from
the black list. Please see the #286 issue in more details.

This patch fixes the bug. After this patch, if users try to update an
already stored credential on the black-listed origin, then update prompt
is shown for the stored credential. However, we still ignores newly
added credentials because the origin is already black-listed.

Refs: #286