Disable AImageReader for ARM64 build on P-OS(and later) (#301)
Currently, 64bit release binary causes a crash when playing video on
some web site such as YouTube. It is an known issue o Android and hasn't
been fixed yet[1].

So, this patch simply disables AImageReader feature for ARM64 build on
P-OS(and later) as a temporary workaround until we have some fix for the
issue.

[1] https://crbug.com/977583

Refs: #300