Refactor SecurityLevelChecker (#308)
Currently, when SecurityLevelChecker is initiated, the last security
level is unknown. The value is updated after first listener is called.
It requires that client code should use static cache by force.
So, this patch fixes the problem. After this patch, the last security
level is updated when the object is created. Also, client's listener
is always called once when adding the listener. Then, we can remove the
static cache in PasswordExtension.

Refs: none