Implement metadata check for subresource filter (#354)
Currently, we trigger to download the subresource filter fully once
every two weeks even if they already have it. It works anyway but it's
a huge waste of data. So, this patch makes it check the filter's
metadata(e.g. version, size, and so on) first and then actually request
to download it if there was a metadata change.

Refs: #348