Implement compressed filter extractor (#357)
To update the adblock filter, the current implementation requests to
download the uncompressed adblock filter from the remote server. It
works anyway but it requires much data consumtion.
So, this patch implements a zip extractor that can extract the
compressed filter. After this patch, we can download the compressed
filter from the server and then extract it on user's device. It also
reduces the data consumption(for downloading filter) to a quarter.

Refs: #348
Co-authored-by: Jinho Bang <zino@chromium.org>