Fix a bug that the home button is sometimes disappeared from top bar (#390)
The home button next to URL bar is sometimes disappeared. This is 
because the upstream code assumed that the home button is on the bottom
toolbar if the bottom toolbar and specific variation are enabled. So,
this patch makes the `isHomeButtonOnBottom()` return always `false` in
order to keep to show the home button next to URL bar.

Bug: #385