Make the secure dns noti to head-up noti only (#409)
When SecureDNS is enabled, our browser shows head-up notification and
then the notification is kept on notification bar. It is general
behavior but it annoy users. If the users felt uncomfortable at this
time, they might block the notification on system settings. And then
they might no longer receive notifications eventually. To resolve this
issue, we only allow head-up notification and make it not kept on
notification bar.

Refs: #406