Make the Secure DNS noti show only once (#412)
Currently, the Secure DNS noti is shown whenever our browser launches if
the option is enabled. It is informative but a little bit annoying. So,
this patch make the Secure DNS noti show only once if the option is
enabled. Once noti was shown, even if restarting our browser again, the
noti is no longer shown.

Refs: #406