Implement HSTS to enforce https (#416)
This patch means implementing HSTS(HTTP Strict Transport Security) which
uses only HTTPS connections. It helps to protect websites against
MITM attacks and improve the usability of the secure DNS feature.  When
using a browser, the users often omit protocols such as HTTP and HTTPS.
In this case, the pre-loaded list is used to force the HTTPS protocol.
Then, the user can use the secure DNS feature as expected.

Refs: #417
Co-authored-by: Jinho Bang <zino@chromium.org>