Store application version name (#420)
Currently, we are not saving our app version name. It is necessary to
store the version in order to process something as per the app update in
the future. So this patch implements it.

This patch checks if there is a preference value stored in the previous
version to detect the first install.

Refs: #404