Fix a bug that the SecureDNS head-up noti doesn't disappear on N- OS (#460)
The setTimeoutAfter() is not working on N- OS. So, the following code is
a workaround to remove the notification after 3 seconds. It works pretty
well, but it may not be canceled if the app exits before the post
delayed task is not executed. Using the AlarmManager may be better, but
it's not worth the effort, so simply use postDelayed().

Refs: #458