Fix a bug that BrowserLock starts after authentication for SecureLogin (#688)
The keyguard authentication runs as a separated system activity from our
application. So, while the keyguard authentication is in progress, our
application enters the background state. Then, after authenticaiton, the
foreground event is fired again, and then it triggers the browser lock.
So, this patch fixes the problem to pause the browser lock during the
keyguard authentication.

Refs: #687