Implement KeyguardBackend and use it (#730)
This patch implements the new KeyguardBackend. In case of the similar
existing KeyguardActivity, the mediator activity is provided between the
authentication requester and the authenticator. However, there is a bug
that the activity would be finished due to unexpected activity life
cycle. Especially, it is reproducible by ChromeLauncherActivity and then
it eventually causes BrowserLock broken.

Refs: #712