Reduce the touch bounds to prevent unexpected gesture detection (#733)
Before this patch, the touch bounds is the same with the target view's
drawing bounds. So, it causes the media gestures to conflict with
opening status bar or navigation bar. To prevent the unexpected gesture
detection, this patch reduces the touch bounds. After this patch, the
bounds is reduced to 90%.

Refs: #725
Co-authored-by: Jinho Bang <zino@chromium.org>