Reduce mojo calls on media event dispatcher (#743)
Currently, we use mojo IPC to hook media events in Blink. But there are
unknown memory leaks. We are not sure but we suspect the following
things:
  - Too many objects are being created
  - Use general mojom type instead of Blink mojom type
So, this patch reduces the mojo bindings and calls by introducing
singleton event dispatcher in Blink and replace the general mojom type
with Blink mojom type.

Refs: #742