Fix a bug the upstream toolbar appears for a while (#783)
This patch means fixing a bug the upstream toolbar appears for a while. It's
regression after #773. At this patch, we moved creating toolbar buttons in
`buttonInitializeWithNative()` to set listener `mTabSwitcherListener`. This
changes the timing for `viewGroup.removeAllViews()`, it turns out the upstream
toolbar appears for a while. So, we just do set `mTabSwitcherListener` on
`buttonInitializeWithNative()`.

Refs: #782