Reimplement subresource filter using JNI instead of mojom IPC (#797)
This patch fixes the issue that the adblock doesn't work after app
update. We still don't know the exact root cause, but it might be
because of introducing `monochrome` in 4.0x (upgraded from 3.02).
This patch fixes the issue trying to update the filter ruleset by force
in a case of app update. To do it, this patch reimplements the
subresource filter module using JNI instead of mojom IPC (because the
app update detection logic is written using Android Java).

Refs: #796