Remove setFilterVersion() method in VersionInfo (#817)
The current implementation sets the version in the FilterLoader when the
ruleset filter is loadable. However, even if the ruleset installation
fails in native side, the version might be displayed as if it were
normal. So, this patch removes the setFilterVersion() and makes it call
the getVersion() in FilterLoader when it needs to display the version.

Refs: #803