Disable bytecode_rewriter_target for androidx_fragment (#1075)
This patch disables bytecode_rewriter_target flag for androidx_fragment
aar. According to the upstream[1], this flag modify the return type of
AndroidX's Fragment.getActivity() method from FragmentActivity to
Activity. It causes the problem[2] in android biometricx aar which used
for our secure login feature. As a stopgap, disable this flag.

[1] https://source.chromium.org/chromium/chromium/src/+/master:docs/ui/android/bytecode_rewriting.md
[2] https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:biometric/biometric/src/main/java/androidx/biometric/BiometricFragment.java;l=262

Refs: #1066