Fix a release build error (#1079)
This patch fixes a release build error about #1075 regression. As the
bytecode_rewriter_target flag is turned off, the code that uses it
should be modified. So, we should manually build the jar file[1] and
then replace the existing one. Additionally, this jar file seems to have
been removed from the current upstream.

[1] third_party/android_deps/local_modifications/androidx_preference_preference/androidx_preference_preference_java.jar

Refs: #1075, #1078